requirements.txt
WTForms~=3.1.2
SQLAlchemy~=2.0.31
Werkzeug~=3.0.3
Flask~=3.0.3
Flask-Login~=0.6.3
Flask-WTF~=1.2.1
python-dotenv~=1.0.1
matplotlib~=3.9.0
Unidecode~=1.3.8
email_validator~=2.2.0